SRC/requirements.txt
nicegui
colorama
PyYaml
